PLN-26: Add questions funcionality

+++ routes/quizzes.js
@@ -1,19 +1,31 @@
 var express = require('express');
 var router = express.Router();
 const { isLoggedIn } = require('../lib/auth');
+const questionHelper = require('../lib/questionHelper');
 const quizHelper = require('../lib/quizHelper');
 //const databaseHelper = require('../lib/databaseHelper');
 
 router.get('/edit', isLoggedIn, async (req, res) => {
     let quizId = req.query.quizId;
     const quiz = await quizHelper.getQuiz(quizId);
-    console.log(quiz);
-    if(quiz != null){
-        res.render('editquiz', {quiz});
+    console.log(quiz.userId);
+    if(quiz.userId == req.user.id){
+
+        const quizQuestions = await questionHelper.getQuestions(quizId);
+        console.log(quizQuestions);
+        console.log(quiz);
+        if(quiz != null){
+            res.render('editquiz', {quiz, quizQuestions});
+        }else{
+            req.flash('message', 'Algo ha ido mal!');
+            res.redirect('/profile');
+        }
+
     }else{
         req.flash('message', 'Algo ha ido mal!');
         res.redirect('/profile');
     }
+    
     
 })
 
@@ -33,12 +45,18 @@ router.post('/add', isLoggedIn, async (req, res) => {
 router.get('/delete', isLoggedIn, async (req, res) => {
     let quizId = req.query.quizId;
     console.log(quizId);
-    if(await quizHelper.deleteQuiz(quizId)){
-        req.flash('success', 'Quiz borrado');
+    if(quiz.userId == req.user.id){
+        if(await quizHelper.deleteQuiz(quizId)){
+            req.flash('success', 'Quiz borrado');
+        }else{
+            req.flash('message', 'Algo ha ido mal!');
+        }
+        res.redirect('/profile');
     }else{
         req.flash('message', 'Algo ha ido mal!');
+        res.redirect('/profile');
     }
-    res.redirect('/profile');
+    
 })
